pyproject.toml
[tool.poetry]
name = "nlp-flashcard-project"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
numpy = "^1.22.3"
transformers = "^4.17.0"
torch = "^1.11.0"
datasets = "^1.18.4"
faiss-cpu = "^1.7.2"
python-dotenv = "^0.19.2"
elasticsearch = "^8.1.0"
gradio = {extras = ["Jinja2"], version = "^2.9.0"}
Jinja2 = "^3.1.1"
coloredlogs = "^15.0.1"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
autopep8 = "^1.6.0"
mypy = "^0.941"
ipykernel = "^6.11.0"

[tool.mypy]
no_implicit_optional=true

[[tool.mypy.overrides]]
module = [
    "transformers",
    "datasets",
]
ignore_missing_imports = true


[tool.isort]
profile = "black"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
